+++ packages/engine/test/Actions/Validate.test.js
@@ -98,17 +98,27 @@ test('Validate required field', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    error: {
+      action: {
+        id: 'validate',
+        type: 'Validate',
+      },
+      error: {
+        error: new Error('Your input has 1 validation error.'),
+        index: 0,
+        type: 'Validate',
+      },
+    },
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
         error: new Error('Your input has 1 validation error.'),
       },
-    ],
-    success: false,
-    timestamp: {
-      date: 0,
     },
+    success: false,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
   expect(text1.validationEval.output).toEqual({
     errors: ['This field is required'],
@@ -134,17 +144,16 @@ test('Validate required field', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
         response: undefined,
       },
-    ],
-    success: true,
-    timestamp: {
-      date: 0,
     },
+    success: true,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
   expect(text1.validationEval.output).toEqual({
     errors: [],
@@ -227,17 +236,27 @@ test('Validate all fields', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    error: {
+      action: {
+        id: 'validate',
+        type: 'Validate',
+      },
+      error: {
+        error: new Error('Your input has 2 validation errors.'),
+        index: 0,
+        type: 'Validate',
+      },
+    },
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
         error: new Error('Your input has 2 validation errors.'),
       },
-    ],
-    success: false,
-    timestamp: {
-      date: 0,
     },
+    success: false,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
   expect(text1.validationEval.output).toEqual({
     errors: ['text1 does not match pattern "text1"'],
@@ -268,17 +287,27 @@ test('Validate all fields', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    error: {
+      action: {
+        id: 'validate',
+        type: 'Validate',
+      },
+      error: {
+        error: new Error('Your input has 1 validation error.'),
+        index: 0,
+        type: 'Validate',
+      },
+    },
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
         error: new Error('Your input has 1 validation error.'),
       },
-    ],
-    success: false,
-    timestamp: {
-      date: 0,
     },
+    success: false,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
   expect(text1.validationEval.output).toEqual({
     errors: [],
@@ -309,17 +338,16 @@ test('Validate all fields', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
         response: undefined,
       },
-    ],
-    success: true,
-    timestamp: {
-      date: 0,
     },
+    success: true,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
   expect(text1.validationEval.output).toEqual({
     errors: [],
@@ -406,17 +434,28 @@ test('Validate only one field', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    error: {
+      action: {
+        id: 'validate',
+        params: 'text1',
+        type: 'Validate',
+      },
+      error: {
+        error: new Error('Your input has 1 validation error.'),
+        index: 0,
+        type: 'Validate',
+      },
+    },
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
         error: new Error('Your input has 1 validation error.'),
       },
-    ],
-    success: false,
-    timestamp: {
-      date: 0,
     },
+    success: false,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
   expect(text1.validationEval.output).toEqual({
     errors: ['text1 does not match pattern "text1"'],
@@ -447,17 +486,16 @@ test('Validate only one field', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
         response: undefined,
       },
-    ],
-    success: true,
-    timestamp: {
-      date: 0,
     },
+    success: true,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
   expect(text1.validationEval.output).toEqual({
     errors: [],
@@ -569,17 +607,28 @@ test('Validate list of fields', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    error: {
+      action: {
+        id: 'validate',
+        params: ['text1', 'text2'],
+        type: 'Validate',
+      },
+      error: {
+        error: new Error('Your input has 1 validation error.'),
+        index: 0,
+        type: 'Validate',
+      },
+    },
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
         error: new Error('Your input has 1 validation error.'),
       },
-    ],
-    success: false,
-    timestamp: {
-      date: 0,
     },
+    success: false,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
   expect(displayMessage.mock.calls).toMatchInlineSnapshot(`
     Array [
@@ -615,17 +664,16 @@ test('Validate list of fields', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
         response: undefined,
       },
-    ],
-    success: true,
-    timestamp: {
-      date: 0,
     },
+    success: true,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
   expect(text3.validationEval.output).toEqual({
     errors: ['text3 does not match pattern "text3"'],
@@ -674,17 +722,30 @@ test('Invalid Validate params', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    error: {
+      action: {
+        id: 'validate',
+        params: {
+          invalid: true,
+        },
+        type: 'Validate',
+      },
+      error: {
+        error: new Error('Invalid validate params.'),
+        index: 0,
+        type: 'Validate',
+      },
+    },
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
         error: new Error('Invalid validate params.'),
       },
-    ],
-    success: false,
-    timestamp: {
-      date: 0,
     },
+    success: false,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
   expect(displayMessage.mock.calls).toMatchInlineSnapshot(`
     Array [
@@ -757,15 +818,14 @@ test('Validate does not fail on warnings', async () => {
     blockId: 'button',
     event: undefined,
     eventName: 'onClick',
-    responses: [
-      {
-        actionId: 'validate',
-        actionType: 'Validate',
+    responses: {
+      validate: {
+        type: 'Validate',
+        index: 0,
       },
-    ],
-    success: true,
-    timestamp: {
-      date: 0,
     },
+    success: true,
+    startTimestamp: { date: 0 },
+    endTimestamp: { date: 0 },
   });
 });