+++ packages/operators/src/web/index.js
@@ -15,10 +15,12 @@
 */
 
 export default {
+  _actions: 'web/actions',
   _event_log: 'web/event_log',
   _base64: 'web/base64',
   _format: 'web/format',
   _list_contexts: 'web/list_contexts',
+  _js: 'web/js',
   _media: 'web/media',
   _menu: 'web/menu',
   _request_details: 'web/request_details',